requirements.txt
redis
requests
uvicorn
fastapi
pymysql
rsa
pyOpenSSL
pycryptodome